resolved problems with duplicating services and models
