+++ backend/Squirrel.Core/Squirrel.Core.WebAPI/appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "SquirrelCoreDBConnection": "Server=localhost;Database=SquirrelCoreDB;Trusted_Connection=True;"
+    "SquirrelCoreDBConnection": "Server=(localdb)\\MSSQLLocalDB;Database=SquirrelCoreDB;Trusted_Connection=True;"
   },
   "MongoDatabase": {
     "ConnectionString": "",
