--- frontend/src/app/models/auth/token-dto.ts
@@ -1,4 +1,0 @@
-export interface TokenDto {
-    accessToken: string;
-    refreshToken: string;
-}
