+++ frontend/src/app/models/auth/user-auth-dto.ts
@@ -1,8 +1,8 @@
 import { UserDto } from '../user/user-dto';
 
-import { TokenDto } from './token-dto';
+import { AccessTokenDto } from './access-token-dto';
 
 export interface UserAuthDto {
     user: UserDto;
-    token: TokenDto;
+    token: AccessTokenDto;
 }
