--- frontend/src/app/models/auth/user-login-dto.ts
@@ -1,4 +1,0 @@
-export interface UserLoginDto {
-    email: string;
-    password: string;
-}
