+++ frontend/src/app/modules/authentication/login/login.component.html
@@ -5,11 +5,7 @@
         <div class="additional-sing-in">
             <app-button class="margin-top" text="Sign In with GitHub" variant="outline-primary" width="100%">
             </app-button>
-            <app-button
-                class="margin-top"
-                id="signInGoogle"
-                text="Sign In with Google"
-                variant="outline-primary"
+            <app-button class="margin-top" id="signInGoogle" text="Sign In with Google" variant="outline-primary"
                 width="100%">
             </app-button>
         </div>
@@ -17,27 +13,18 @@
         <app-barrier [text]="'or'"></app-barrier>
 
         <form class="login-form" (ngSubmit)="loginForm.valid && login()">
-            <app-input [formControl]="$any(loginForm.controls['email'])" [label]="'email'"> </app-input>
-
-            <app-input [formControl]="$any(loginForm.controls['email'])"
-                       [label]="'email'"
-                       placeholder="Enter your Email">
+            <app-input [formControl]="$any(loginForm.controls['email'])" [label]="'email'"
+                placeholder="Enter your Email">
             </app-input>
 
             <div class="input-wrapper">
-                <app-input [formControl]="$any(loginForm.controls['password'])"
-                           [label]="'password'"
-                           [type]="'password'"
-                           placeholder="Enter your Password">
+                <app-input [formControl]="$any(loginForm.controls['password'])" [label]="'password'" [type]="'password'"
+                    placeholder="Enter your Password">
                 </app-input>
                 <a href="">Forget password</a>
             </div>
 
-            <app-button
-                class="margin-top"
-                [isDisabled]="loginForm.invalid"
-                text="Sing In"
-                variant="filled"
+            <app-button class="margin-top" [isDisabled]="loginForm.invalid" text="Sing In" variant="filled"
                 width="100%">
             </app-button>
         </form>
